+++ Tests/XcodeTests/UIKitProject/UIKitProject/XibViewController.swift
@@ -2,9 +2,12 @@ import UIKit
 
 class XibViewController: UIViewController {
     @IBOutlet weak var button: UIButton!
+    @IBOutlet weak var unusedOutlet: UILabel!
     @IBAction func click(_ sender: Any) {}
     @IBAction func clickFromSubclass(_ sender: Any) {}
+    @IBAction func unusedAction(_ sender: Any) {}
     @IBInspectable var controllerProperty: UIColor?
+    @IBInspectable var unusedInspectable: CGFloat = 0
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -21,6 +24,7 @@ class XibView: UIView {
 }
 
 extension UIView {
+    // Referenced via XIB (used in userDefinedRuntimeAttributes)
     @IBInspectable var customBorderColor: UIColor? {
         get {
             if let borderColor = layer.borderColor {
@@ -32,5 +36,11 @@ extension UIView {
         set {
             layer.borderColor = newValue?.cgColor
         }
+    }
+
+    // Unreferenced - not used in any XIB
+    @IBInspectable var unusedExtensionInspectable: CGFloat {
+        get { 0 }
+        set { _ = newValue }
     }
 }