+++ Tests/XcodeTests/UIKitProject/UIKitProject/XibViewController2.swift
@@ -1,8 +1,13 @@
 import UIKit
 
 class XibViewController2Base: UIViewController {
-    @IBOutlet weak var button: UIButton!
+    // Referenced via XIB (connected)
     @IBAction func clickFromSubclass(_ sender: Any) {}
+
+    // Unreferenced - not connected in XIB
+    @IBOutlet weak var button: UIButton!
+    @IBOutlet weak var unusedBaseOutlet: UILabel!
+    @IBAction func unusedBaseAction(_ sender: Any) {}
 }
 
 class XibViewController2Subclass: XibViewController2Base {}